Create LocateIPAddress - 1st Commit
Current progress of the LocateIPAddress project. Very little exception handling is added, and the API's are spotty. IPv4 regex checks work fine, but IPv6 checks do not. Lastly, the OutputtoJson parameter isn't active yet, but will wrap up the outputs to output to a variable and tie that to the option.